Add code to edit and delete views to ensure only the user who created link can edit and delete

+++ links/views.py
@@ -43,39 +43,46 @@ def add_link(request):
 
 @login_required
 def edit_link(request, url_links_id):
-    item = get_object_or_404(url_links, id=url_links_id)
+    link = get_object_or_404(url_links, id=url_links_id)
     
+    if request.user == link.user:
 
-    if request.method == 'POST':
-        form = LinkForm(request.POST, instance=item)
-        form_url = request.POST.get('link')
-        url_pass=''
-        try:
-            requests.get(form_url)
-            url_pass=True
-        except requests.exceptions.ConnectionError:
-            messages.warning(request, 'Invalid URL entered')
-            url_pass=False
-        if form.is_valid() and url_pass==True:
-            edit_form = form.save(commit=False)
-            edit_form.user = request.user
-            edit_form.save()
-            return redirect('links-home')
-        
-    form = LinkForm(instance=item)
-    context = {
-        'form': form
-        }
-    return render(request, 'edit_item.html', context)
+        if request.method == 'POST':
+            form = LinkForm(request.POST, instance=link)
+            form_url = request.POST.get('link')
+            url_pass=''
+            try:
+                requests.get(form_url)
+                url_pass=True
+            except requests.exceptions.ConnectionError:
+                messages.warning(request, 'Invalid URL entered')
+                url_pass=False
+            if form.is_valid() and url_pass==True:
+                edit_form = form.save(commit=False)
+                edit_form.user = request.user
+                edit_form.save()
+                return redirect('links-home')
+            
+        form = LinkForm(instance=link)
+        context = {
+            'form': form
+            }
+        return render(request, 'edit_item.html', context)
+    else:
+        return redirect('links-home')
 
 @login_required
 def delete_link(request, url_links_id):
     link = get_object_or_404(url_links, id=url_links_id)
     
-    if request.method=='POST' and request.user == link.user:
-        link.delete()
+    if request.user == link.user:
+
+        if request.method=='POST' and request.user == link.user:
+            link.delete()
+            return redirect('links-home')
+    
+        return render(request, 'confirm_delete.html',{'link':link})
+    else:
         return redirect('links-home')
-   
-    return render(request, 'confirm_delete.html',{'link':link})
